fix: SPM 에서 swiftlint 삭제
