+++ ios/AirBnb/.swiftlint.yml
@@ -41,7 +41,7 @@ type_name:
   excluded: iPhone # 제외할 문자열 값 사용
 identifier_name:
   min_length: # min_length에서
-    error: 3 # 에러만 적용
+    error: 1 # 에러만 적용
   excluded: # 제외할 문자열 목록 사용
     - id
     - URL
