+++ ios/AirBnb/AirBnb/SceneDelegate.swift
@@ -14,6 +14,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     guard let scene = (scene as? UIWindowScene) else {
       return
     }
+
     self.window = UIWindow(windowScene: scene)
 
     self.window?.rootViewController = ViewController()
