+++ ios/AirBnb/Toolbox/Sources/Toolbox/Toolbox.swift
@@ -1,0 +1,5 @@
+public struct Toolbox {
+  public private(set) var text = "Hello, World!"
+
+  public init() {}
+}
