+++ ios/AirBnb/Toolbox/Tests/ToolboxTests/ToolboxTests.swift
@@ -1,0 +1,11 @@
+import XCTest
+@testable import Toolbox
+
+final class ToolboxTests: XCTestCase {
+  func testExample() throws {
+    // This is an example of a functional test case.
+    // Use XCTAssert and related functions to verify your tests produce the correct
+    // results.
+    XCTAssertEqual(Toolbox().text, "Hello, World!")
+  }
+}
