--- ios/BuildTools/Sources/BuildTools/BuildTools.swift
@@ -1,6 +1,0 @@
-public struct BuildTools {
-    public private(set) var text = "Hello, World!"
-
-    public init() {
-    }
-}
