--- ios/BuildTools/Tests/BuildToolsTests/BuildToolsTests.swift
@@ -1,11 +1,0 @@
-import XCTest
-@testable import BuildTools
-
-final class BuildToolsTests: XCTestCase {
-    func testExample() throws {
-        // This is an example of a functional test case.
-        // Use XCTAssert and related functions to verify your tests produce the correct
-        // results.
-        XCTAssertEqual(BuildTools().text, "Hello, World!")
-    }
-}
